Added API link to menu, organized API folder, changed dockerfile, readme
