+++ README.md
@@ -62,7 +62,7 @@ Then:
 
 ```
 pip install -r requirements.txt
-pip install --upgrade git+git://github.com/Yelp/swagger_spec_validator
+pip install --upgrade git+git://github.com/Yelp/swagger_spec_validator git+https://github.com/pierotofy/django-knockout
 chmod +x start.sh && ./start.sh
 ```
