+++ app/templates/app/logged_in_base.html
@@ -263,6 +263,9 @@
                 </li>
                 {% endif %}
                 <li>
+                    <a href="/api/"><i class="fa fa-book fa-fw"></i> {% trans 'API' %}</a>
+                </li>
+                <li>
                     <a href="#"><i class="fa fa-life-saver fa-fw"></i> {% trans 'Documentation' %}</a>
                 </li>
             </ul>
