+++ app/urls.py
@@ -1,28 +1,10 @@
 from django.conf.urls import url, include
-from . import views, models
-from django.contrib.auth.models import User
-from rest_framework import routers, serializers, viewsets
-
-class ProjectSerializer(serializers.HyperlinkedModelSerializer):
-    owner = serializers.PrimaryKeyRelatedField(queryset=User.objects.all())
-
-    class Meta:
-        model = models.Project
-
-
-class ProjectViewSet(viewsets.ModelViewSet):
-    queryset = models.Project.objects.all()
-    serializer_class = ProjectSerializer
-
-# Routers provide an easy way of automatically determining the URL conf.
-router = routers.DefaultRouter()
-router.register(r'projects', ProjectViewSet)
+from . import views
 
 urlpatterns = [
     url(r'^$', views.index, name='index'),
     url(r'^dashboard/$', views.dashboard, name='dashboard'),
     url(r'^processingnode/([\d]+)/$', views.processing_node, name='processing_node'),
 
-    url(r'^api/', include(router.urls)),
-    url(r'^api/auth/', include('rest_framework.urls')),
+    url(r'^api/', include("app.api.urls")),
 ]
